src/agent/pyproject.toml
[project]
name = "investment-copilot-agent"
version = "0.1.0"
description = "Investment Copilot Agent for Databricks Model Serving"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.81.0",
    "mlflow>=2.12.0",
    "langchain-core>=0.3.0",
    "langchain-databricks>=0.1.0",
    "langgraph>=0.2.0",
]
